% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{as.dictionary}
\alias{as.dictionary}
\alias{as.dictionary.list}
\alias{as.dictionary.default}
\alias{as.dictionary.data.frame}
\title{Coerce to a \link[=new_dictionary]{LabelDictionary} class object}
\usage{
as.dictionary(.data, ...)

\method{as.dictionary}{list}(.data)

\method{as.dictionary}{default}(...)

\method{as.dictionary}{data.frame}(.data, translation, col_old, col_new,
  ordering = "row")
}
\arguments{
\item{.data}{An object holding the translations. \code{.data} can be of the
following data types:
\itemize{
\item \emph{named list}: A named list object, where each list element is a translation
(a named character vector)
\item \emph{data.frame}: A data.frame holding one or more column pairs,
where each column pair consists of
one column holding the original variable values and
a second column holding the new labels,
which should be assigned to the original values.
}}

\item{...}{Various arguments, depending on the data type of \code{.data}.}

\item{translation}{A character vector holding the names of all translations}

\item{col_old}{A character vector (same length as \code{translation}) holding the
names of the columns in the data.frame (in the argument \code{.data}) which hold
the original variable values.
These columns can be of any type: \code{character}, \code{logical}, \code{numerical} or \code{factor}.}

\item{col_new}{A character vector (same length as \code{translation}) holding the
names of the columns in the data.frame (in the argument \code{.data}) which hold
the new labels, which should be assigend to the original values.
These columns can be \code{character} vectors or \code{factors} with character labels.}
}
\value{
A new LabelDictionary class object holding the passed in translations.
}
\description{
This function allows two types of arguments:
\itemize{
\item \emph{named list}: A named list object holding the translations.
\item \emph{data.frame}: A data.frame with one ore more column pairs. Each column
pair consists of a column holding the original values, which should be relapced,
and a second charachter column holding the new labels which should be
assigned to the original values. Use the arguments \code{col_old} and \code{col_new}
in order to define which columns are holding original values and which
columns hold the new labels. The names of the resulting translations
are defined by a character vector given in argument \code{translation}.
Furthermore, each translation can have a different ordering which can be
configured by a character vector given in argument \code{ordering}.
}
}
\section{Translations}{

A \emph{translation} is a \emph{named character vector} of non zero length.
This named character vector defines
which labels (of type character) should be assigned to which values
(can be of type character, logical or numeric)
(e.g. the translation \code{c("0" = "urban", "1" = "rural")} assigns the label
\code{"urban"} to the value \code{0} and \code{"rural"} to the value \code{1}, for example the
variable \code{x = c(0, 0, 1)} is translated to \code{x_new = c("urban", "urban", "rural")}).
Therefore, a translation (named character vector) contains the following information:
\itemize{
\item The \emph{names} of the character vector entries correspond to the
\emph{original variable levels}.
Variables of types \code{numeric} or \code{logical} are turned automatically into a
character vector (e.g. \code{0} and \code{1} are treated like \code{"0"} and \code{"1"}).
\item The \emph{entries} (character strings) of the character vector correspond to
the new \emph{labels}, which will be assigned to the original variable levels.
It is also allowed to have missing labels (\code{NA}s).
In this case, the original values are mappend to missing values.
}

The function \code{\link[=lama_translate]{lama_translate()}} is used in order to apply a translation on a variable.
The resulting vector with the assigned labels can be of the following types:
\itemize{
\item \emph{character}: An unordered vector holding the new character labels.
\item \emph{factor} with character levels: An ordered vector holding the new character labels.
}

The original variable can be of the following types:
\itemize{
\item \emph{character} vector: This is the simplest case. The character values
will replaced by the corresponding labels.
\item \emph{numeric} or \emph{logical} vector: Vectors of type \emph{numeric} or \emph{logical}
will be turned
into \emph{character} vectors automatically before the translation process and
then simply processed like in the \emph{character} case.
Therefore, it is sufficient to define the translation mapping for the
\emph{character} case, since it also covers the \emph{numeric} and \emph{logical} case.
\item \emph{factor} vector with levels of any type: When translating factor variables
one can decide whether or not to keep the original ordering. Like in the
other cases the levels of the factor variable will always be turned into
character strings before the translation process.
}
}

\section{LabelDictionary class objects}{

Each \emph{LabelDictionary} class object can contain multiple \emph{translations},
each with a unique name under which the translation can be found.
The function \code{\link[=lama_translate]{lama_translate()}} uses a LabelDictionary class object
to translate a normal \code{vector} or to translate one or more columns in a
\code{data.frame}.
Sometimes it may be necessary to have different translations
for the same variable, in this case it is best to have multiple
translations with different names
(e.g. \code{area_short = c("0" = "urb", "1" = "rur")} and
\code{area = c("0" = "urban", "1" = "rural")}).
}

