% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_rename.R
\name{check_rename}
\alias{check_rename}
\title{Function that checks the passed in arguments for \code{\link[=lama_rename]{lama_rename()}} and \code{\link[=lama_rename_]{lama_rename_()}}}
\usage{
check_rename(.data, old, new, err_handler)
}
\arguments{
\item{.data}{A \link[=new_lama_dictionary]{lama_dictionary} object, holding the
variable translations}

\item{old}{A character vector holding the names of the variable
translations, that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}

\item{err_handler}{A error handling function}
}
\description{
Function that checks the passed in arguments for \code{\link[=lama_rename]{lama_rename()}} and \code{\link[=lama_rename_]{lama_rename_()}}
}
