% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{contains_na_escape}
\alias{contains_na_escape}
\title{Check if a character vector contains NA replacement strings}
\usage{
contains_na_escape(x)
}
\arguments{
\item{x}{A character vector that should be checked.}
}
\value{
\code{TRUE} if the vector contains NA replacement strings. \code{FALSE} else.
}
\description{
Check if a character vector contains NA replacement strings
}
