% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dictionary_to_yaml}
\alias{dictionary_to_yaml}
\title{Transform data structure from [LabelDictionary] class input format to the yaml format}
\usage{
dictionary_to_yaml(data)
}
\description{
In the [LabelDictionary] class object the data has the structure
vars (named list) > translations (named character vector)
This structure is transformed to the yaml file structure
vars (named list) > translations (named list)
}
