% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{escape_to_na}
\alias{escape_to_na}
\title{Replace \code{"NA_"} by \code{NA}}
\usage{
escape_to_na(x)
}
\arguments{
\item{x}{A character vector that should be modified.}
}
\value{
A character vector, where the NA replacement strings are replaced by \code{NA}s.
}
\description{
Replace \code{"NA_"} by \code{NA}
}
