% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{is.dictionary}
\alias{is.dictionary}
\title{Check if an object is a \link[=new_dictionary]{LabelDictionary} class object}
\usage{
is.dictionary(x)
}
\arguments{
\item{x}{The object in question}
}
\value{
\code{TRUE} if the object is a \link[=new_dictionary]{LabelDictionary} class object, \code{FALSE} otherwise.
}
\description{
Check if an object is a \link[=new_dictionary]{LabelDictionary} class object
}
\seealso{
\code{\link[=as.dictionary]{as.dictionary()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=translate]{translate()}}, \code{\link[=read_directory]{read_directory()}}, \code{\link[=write_directory]{write_directory()}}, \code{\link[=select]{select()}}, \code{\link[=rename]{rename()}}, \code{\link[=mutate]{mutate()}}, \code{\link[=merge]{merge()}}
}
