% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_lexicon.R, R/merge_lexica.R
\name{new_lexicon}
\alias{new_lexicon}
\alias{select}
\alias{select.LabelLexicon}
\alias{rename}
\alias{rename.LabelLexicon}
\alias{mutate}
\alias{mutate.LabelLexicon}
\alias{print.LabelLexicon}
\alias{merge_lexicas}
\title{Constructor function of the [LabelLexicon] class}
\usage{
new_lexicon(translations = NULL)

select(.data, ...)

\method{select}{LabelLexicon}(.data, translation)

rename(.data, ...)

\method{rename}{LabelLexicon}(.data, old, new)

mutate(.data, ...)

\method{mutate}{LabelLexicon}(.data, translation, value)

\method{print}{LabelLexicon}(x)

merge_lexicas(..., show_warnings = TRUE)

merge_lexicas(..., show_warnings = TRUE)
}
\arguments{
\item{translations}{A named list object holding your label translations. Each
list item must be a named character vector, which represent the translations replacing
the old labels of the variables by new labels.
The item names of the character vectors correspond to the old variable
labels which should be relabed and
the values are the new labels that should be used. The ordering of the 
character vectors will be used as the ordering of the new labels.}

\item{.data}{A [LabelLexicon] object, holding the label translations}

\item{...}{Various arguments}

\item{translation}{A character vector holding the names of the translations that
should be picked.}

\item{old}{A character vector holding the names of the translations, that should be renamed.}

\item{new}{A character vector holding the new names of the translations.}

\item{value}{A named character vector holding the new label translation
information. The names of the character vector correspond to the old variabel levels that should be replaced.
The values in the character values are the new labels that should be assigned.}

\item{x}{The [LabelLexicon] class object that should be printed.}

\item{.data}{A [LabelLexicon] object, holding the label translations}

\item{.data}{A [LabelLexicon] object, holding the label translations}

\item{...}{Various arguments old A character vector holding the names of the translations that
should be picked.}

\item{.data}{A [LabelLexicon] object, holding the label translations}

\item{.data}{A [LabelLexicon] object}

\item{...}{Various arguments old A character vector holding the names of the translations that}

\item{.data}{A [LabelLexicon] object}

\item{translation}{The name of the translation that should be altered.
It can also be translation name that does not exist yet.}

\item{...}{Multiple [LabelLexicon] class objects arguments}

\item{...}{Multiple [LabelLexicon] class objects arguments}

\item{show_wanings}{A logical flag that defines, whether warnings should be shown (\code{TRUE}) or not (\code{FALSE})}
}
\value{
A new [LabelLexicon] class object, holding the picked label translations.

A new [LabelLexicon] class object, holding the picked label translations.

A new [LabelLexicon] class object, holding the picked label translations.

The updated [LabelLexicon] class object.

An updated [LabelLexicon] class object.

An updated [LabelLexicon] class object.

The merged [LabelLexicon] class object

The merged [LabelLexicon] class object
}
\description{
An \code{S3 class} which holds the label translations of various variables.

This function picks one or more translations from a [LabelLexicon] class object
and creates a new [LabelLexicon] class object.

This function renamess one or more translations inside of a [LabelLexicon] class object.

This function renamess one or more translations inside of a [LabelLexicon] class object.

This function alters a [LabelLexicon] object. It can be used for altering
or appending a translation to a [LabelLexicon] object.

This function alters a [LabelLexicon] object. It can be used for altering
or appending a translation to a [LabelLexicon] object.

This takes multiple [LabelLexicon] class objects and merges them together into
a single [LabelLexicon] class object.
In case some class objects have entries with the same name, the 
class objects passed in later overwrite the class objects passed in first 
(e.g. in \code{merge_lexicas(x, y, z)}: The lexicon \code{z} overwrites
\code{x} and \code{y}. The lexicon \code{y} overwrites \code{x}).

This takes multiple [LabelLexicon] class objects and merges them together into
a single [LabelLexicon] class object.
In case some class objects have entries with the same name, the 
class objects passed in later overwrite the class objects passed in first 
(e.g. in \code{merge_lexicas(x, y, z)}: The lexicon \code{z} overwrites
\code{x} and \code{y}. The lexicon \code{y} overwrites \code{x}).
}
