% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_check.R
\name{lama_check_length}
\alias{lama_check_length}
\title{Check if an object has the right length}
\usage{
lama_check_length(obj, comp_obj = NULL, comp_len = NULL,
  err_handler = composerr("Error in 'lama_check_length"))
}
\arguments{
\item{obj}{An object that should be checked if it has the right length}

\item{comp_obj}{A object that should be used for comparison}

\item{comp_len}{A numeric vector holding valid length values}

\item{err_handler}{An error handling function, which should be used}
}
\description{
Check if an object has the right length
}
