% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_merge.R
\name{lama_merge}
\alias{lama_merge}
\alias{lama_merge.lama_dictionary}
\title{Merge multiple label lexicas into one}
\usage{
lama_merge(..., show_warnings = TRUE)

\method{lama_merge}{lama_dictionary}(..., show_warnings = TRUE)
}
\arguments{
\item{...}{Two or more \link[=new_dictionary]{lama_dictionary}
class objects, which should be merged together.}

\item{show_warnings}{A logical flag that defines, whether warnings should be
shown (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The merged \link[=new_dictionary]{lama_dictionary} class object
}
\description{
This function takes multiple \link[=new_dictionary]{lama_dictionary} class
objects and merges them together into
a single \link[=new_dictionary]{lama_dictionary} class object.
In case some class objects have entries with the same name, the
class objects passed in later overwrite the class objects passed in first
(e.g. in \code{lama_merge(x, y, z)}: The lexicon \code{z} overwrites
\code{x} and \code{y}. The lexicon \code{y} overwrites \code{x}).
}
\seealso{
\code{\link[=lama_translate]{lama_translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=lama_rename]{lama_rename()}}, \code{\link[=lama_select]{lama_select()}},
\code{\link[=lama_mutate]{lama_mutate()}}, \code{\link[=lama_read]{lama_read()}}, \code{\link[=lama_write]{lama_write()}}
}
