% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_read.R
\name{lama_read}
\alias{lama_read}
\title{Read in a \code{yaml} file holding translations for one or multiple variables}
\usage{
lama_read(yaml_path)
}
\arguments{
\item{yaml_path}{Path to yaml file holding the labels and translations for multiple variables}
}
\value{
A \link[=new_lama_dictionary]{lama_dictionary} class object holding the variable translations defined in
the yaml file
}
\description{
Read in a \code{yaml} file holding translations for one or multiple variables
}
\examples{
  \dontrun{
    dict <- lama_read("my_dictionary.yaml")
  }
}
