% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_translate.R
\name{lama_translate}
\alias{lama_translate}
\alias{lama_translate.data.frame}
\alias{lama_translate_}
\alias{lama_translate_.data.frame}
\title{Assign new labels to a variable of a data.frame}
\usage{
lama_translate(.data, dictionary, ..., keep_order = FALSE)

\method{lama_translate}{data.frame}(.data, dictionary, ...,
  keep_order = FALSE)

lama_translate_(.data, dictionary, translation, col = translation,
  col_new = col, keep_order = FALSE)

\method{lama_translate_}{data.frame}(.data, dictionary, translation,
  col = translation, col_new = col, keep_order = FALSE)
}
\arguments{
\item{.data}{The data.frame object which contains the variable that should be relabelled}

\item{dictionary}{A \link[=new_dictionary]{LamaDictionary} object, holding the translations for various
variables.}

\item{...}{One or more unquoted expressions separated by commas. Use unquoted
arguments that tell which translation should be applied to which column and
which column name the relabelled variable should be assigned to. E.g.
\code{lama_trans(.data, dict, Y1 = TRANS1(X1), Y2 = TRANS2(Y2))}
to apply the translations \code{TRANS1} and \code{TRANS2} to the data.frame
variables \code{X1} and \code{X2} and save the new labelled variables under
the names \code{Y1} and \code{Y2}.
There are also two abbreviation mechanisms available:
The argument assignement \code{FOO(X)} is the same as \code{X = FOO(X)} and
\code{FOO} is an abbreviation for \code{FOO = FOO(FOO)}.}

\item{keep_order}{A boolean vector of length one or the same length as the
number of arguments in \code{...}. If the vector has length one, then the same
configuration is applied to all variable translations. If the vector has
the same length as the number of arguments in \code{...}, then the
to each variable translation there is a corresponding boolean configuration.
If a translated variable in the data.frame is a factor variable,
and the corresponding boolean configuration is set to \code{TRUE}, then the
the order of the original factor variable will be preserved.}

\item{translation}{A character vector holding the names of the variable
translations which
should be used for assigning new labels to the variable. This names must be
a subset of the translation names returned by \code{names(dictionary)}.}

\item{col}{A character vector of the same length as \code{translation} holding
the names of the data.frame columns that
should be relabelled. If omitted, then it will be assumed that the column
names are the same as the given translation names in the argument \code{translation}.}

\item{col_new}{A character vector of the same length as \code{translation} holding
the names under which the relabelled variables should be stored in
the data.frame. If omitted, then it will be assumed that the new column
names are the same as the column names of the original variables.}
}
\value{
An extended data.frame, that has a factor variable holding the
assigned labels.
}
\description{
The functions \code{\link[=lama_translate]{lama_translate()}} and \code{\link[=lama_translate_]{lama_translate_()}} take a data.frame
and convert one or more of its categorical variables
(not necessarely a factor variable) into factor variables with new labels.
The function \code{\link[=lama_translate]{lama_translate()}} uses non-standard evaluation, whereas
\code{\link[=lama_translate_]{lama_translate_()}} is the standard evaluation alternative.
}
