% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_write.R
\name{lama_write}
\alias{lama_write}
\title{Write a \code{yaml} file holding translations for one or multiple variables}
\usage{
lama_write(x, yaml_path)
}
\arguments{
\item{x}{A \link[=new_lama_dictionary]{lama_dictionary} class object holding the variable translations}

\item{yaml_path}{File path, where the yaml file should be saved}
}
\description{
Write a \code{yaml} file holding translations for one or multiple variables
}
\examples{
  \donttest{
    dict <- new_lama_dictionary(results = c(p = "Passed", f = "Failed"))
    path_to_file <- file.path(tempdir(), "my_dictionary.yaml")
    lama_write(dict, path_to_file)
  }
}
