% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lexicon_to_yaml}
\alias{lexicon_to_yaml}
\title{Transform data structure from [LabelLexicon] class input format to the yaml format}
\usage{
lexicon_to_yaml(data)
}
\description{
In the [LabelLexicon] class object the data has the structure
vars (named list) > labels (named character vector)
This structure is transformed to the yaml file structure
vars (named list) > labels (named list)
}
