% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{mutate}
\alias{mutate}
\alias{mutate.LabelDictionary}
\title{Change a variable translation or append a variable translation to an existing \link[=new_dictionary]{LabelDictionary} object}
\usage{
mutate(.data, ...)

\method{mutate}{LabelDictionary}(.data, variable, translation)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object}

\item{...}{Various arguments}

\item{variable}{The name of the variable translation that should be altered.
If can also be variable translation name that does not exist yet.}

\item{translation}{A named character vector holding the new variable
translation that should be assigned to the name given in argument \code{variable}.
The names of the character vector \code{translation} correspond to the
original variable values that should be replaced by the new labels.
The values in the character vector \code{translations} are the labels
that should be assigned to the original values.}
}
\value{
An updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
This function alters a \link[=new_dictionary]{LabelDictionary} object. It can be used for altering
or appending a translation to a \link[=new_dictionary]{LabelDictionary} object.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=select]{select()}}, \code{\link[=rename]{rename()}}, \code{\link[=merge]{merge()}}, \code{\link[=read_directory]{read_directory()}}, \code{\link[=write_directory]{write_directory()}}
}
