% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_dictionary.R
\docType{data}
\name{NA_lama_}
\alias{NA_lama_}
\title{NA replace string}
\format{An object of class \code{character} of length 1.}
\usage{
NA_lama_
}
\description{
In order to replace \code{NA} values in yaml files and in translations
the following character string is used
}
\keyword{datasets}
