% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{print.LabelDictionary}
\alias{print.LabelDictionary}
\title{Print a [LabelDictionary] class object}
\usage{
\method{print}{LabelDictionary}(x)
}
\arguments{
\item{x}{The [LabelDictionary] class object that should be printed.}
}
\description{
Print a [LabelDictionary] class object
}
