% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dictionary.R
\name{read_dictionary}
\alias{read_dictionary}
\title{Read in a \code{yaml} file holding translations for one or multiple variables}
\usage{
read_dictionary(yaml_path)
}
\arguments{
\item{yaml_path}{Path to yaml file holding the labels and translations for multiple variables}
}
\value{
A \link[=new_dictionary]{LabelDictionary} class object holding the variable translations defined in
the yaml file
}
\description{
Read in a \code{yaml} file holding translations for one or multiple variables
}
