% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yaml_file.R
\name{read_lexicon_file}
\alias{read_lexicon_file}
\title{Read in a \code{yaml} file holding label translations for one or multiple variables}
\usage{
read_lexicon_file(yaml_path)
}
\arguments{
\item{yaml_path}{Path to yaml file holding the labels and translations for multiple variables}
}
\value{
A [LabelLexicon] class object holding the label mappings defined in
the yaml file
}
\description{
Read in a \code{yaml} file holding label translations for one or multiple variables
}
