% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{rename}
\alias{rename}
\alias{rename.LabelDictionary}
\title{Rename multiple variable translations in a \link[=new_dictionary]{LabelDictionary} object}
\usage{
rename(.data, ...)

\method{rename}{LabelDictionary}(.data, old, new)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{Various arguments}

\item{old}{A character vector holding the names of the variable translations, that should be renamed.}

\item{new}{A character vector holding the new names of the variable translations.}
}
\value{
The updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
This function renamess one or more variable translations inside of a \link[=new_dictionary]{LabelDictionary} class object.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=select]{select()}}, \code{\link[=mutate]{mutate()}}, \code{\link[=merge]{merge()}}, \code{\link[=read_directory]{read_directory()}}, \code{\link[=write_directory]{write_directory()}}
}
