% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R, R/trans_rename.R
\name{rename_translation}
\alias{rename_translation}
\title{Function that actually performs the renaming of the translations}
\usage{
rename_translation(.data, old, new)

rename_translation(.data, old, new)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the
variable translations}

\item{old}{A character vector holding the names of the variable
translations, that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}

\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the
variable translations}

\item{old}{A character vector holding the names of the variable
translations, that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}
}
\value{
The updated \link[=new_dictionary]{LabelDictionary} class object.

The updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
Function that actually performs the renaming of the translations

Function that actually performs the renaming of the translations
}
