% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{select}
\alias{select}
\alias{select.LabelDictionary}
\title{Pick multiple variable translations and create a new [LabelDictionary] object}
\usage{
select(.data, ...)

\method{select}{LabelDictionary}(.data, variable)
}
\arguments{
\item{.data}{A [LabelDictionary] object, holding the variable translations}

\item{...}{Various arguments}

\item{variable}{A character vector holding the names of the variable translations that
should be picked.}
}
\value{
A new [LabelDictionary] class object, holding the picked variable translations.
}
\description{
This function picks one or more variable translations from a [LabelDictionary] class object
and creates a new [LabelDictionary] class object.
}
