% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stringify}
\alias{stringify}
\title{Stringify a vector into a single character string (\code{'x1', 'x2', ...})}
\usage{
stringify(x)
}
\description{
Stringify a vector into a single character string (\code{'x1', 'x2', ...})
}
