% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_merge.R
\name{trans_merge}
\alias{trans_merge}
\alias{trans_merge.LabelDictionary}
\title{Merge multiple label lexicas into one}
\usage{
trans_merge(.data, x, y, ..., show_warnings = TRUE)

\method{trans_merge}{LabelDictionary}(x, y, ..., show_warnings = TRUE)
}
\arguments{
\item{x}{A \link[=new_dictionary]{LabelDictionary} class object}

\item{y}{A \link[=new_dictionary]{LabelDictionary} class object}

\item{...}{Optional additional multiple \link[=new_dictionary]{LabelDictionary}
class objects}

\item{show_wanings}{A logical flag that defines, whether warnings should be
shown (\code{TRUE}) or not (\code{FALSE})}
}
\value{
The merged \link[=new_dictionary]{LabelDictionary} class object
}
\description{
This function takes multiple \link[=new_dictionary]{LabelDictionary} class
objects and merges them together into
a single \link[=new_dictionary]{LabelDictionary} class object.
In case some class objects have entries with the same name, the
class objects passed in later overwrite the class objects passed in first
(e.g. in \code{trans_merge(x, y, z)}: The lexicon \code{z} overwrites
\code{x} and \code{y}. The lexicon \code{y} overwrites \code{x}).
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=trans_rename]{trans_rename()}}, \code{\link[=tran_select]{tran_select()}},
\code{\link[=trans_set]{trans_set()}}, \code{\link[=read_dictionary]{read_dictionary()}}, \code{\link[=write_dictionary]{write_dictionary()}}
}
