% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R, R/trans_rename.R
\name{trans_rename}
\alias{trans_rename}
\alias{trans_rename_}
\alias{trans_rename_.LabelDictionary}
\title{Rename multiple variable translations in a \link[=new_dictionary]{LabelDictionary} object}
\usage{
trans_rename(.data, ...)

trans_rename_(.data, old, new)

\method{trans_rename_}{LabelDictionary}(.data, old, new)

trans_rename(.data, ...)

trans_rename_(.data, old, new)

\method{trans_rename_}{LabelDictionary}(.data, old, new)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the
variable translations}

\item{...}{One or more unquoted expressions separated by commas. Use named
arguments, e.g. ‘new_name = old_name’, to rename selected variables.}

\item{old}{A character vector holding the names of the variable translations,
that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}

\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{One or more unquoted expressions separated by commas. Use named arguments, e.g. ‘new_name = old_name’, to rename selected variables.}

\item{old}{A character vector holding the names of the variable translations,
that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}
}
\value{
The updated \link[=new_dictionary]{LabelDictionary} class object.

The updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
The functions \code{\link[=trans_rename]{trans_rename()}} and \code{\link[=trans_rename_]{trans_rename_()}}
are used to rename one or more variable translations inside of a
\link[=new_dictionary]{LabelDictionary} class object.
The function \code{\link[=trans_rename]{trans_rename()}} uses non-standard evaluation,
whereas \code{\link[=trans_rename_]{trans_rename_()}} is the standard evaluation alternativ.

The functions \code{\link[=trans_rename]{trans_rename()}} and \code{\link[=trans_rename_]{trans_rename_()}}
are used to rename one or more variable translations inside of a
\link[=new_dictionary]{LabelDictionary} class object.
The function \code{\link[=trans_rename]{trans_rename()}} uses non-standard evaluation,
whereas \code{\link[=trans_rename_]{trans_rename_()}} is the standard evaluation alternativ.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=trans_select]{trans_select()}},
\code{\link[=trans_mutate]{trans_mutate()}}, \code{\link[=trans_merge]{trans_merge()}}, \code{\link[=read_dictionary]{read_dictionary()}}, \code{\link[=write_dictionary]{write_dictionary()}}

\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=trans_select]{trans_select()}}, \code{\link[=trans_set]{trans_set()}},
\code{\link[=trans_merge]{trans_merge()}}, \code{\link[=read_dictionary]{read_dictionary()}}, \code{\link[=write_dictionary]{write_dictionary()}}
}
