% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_select.R
\name{trans_select}
\alias{trans_select}
\alias{trans_select_}
\alias{trans_select_.LabelDictionary}
\title{Select multiple variable translations and create a new \link[=new_dictionary]{LabelDictionary} object}
\usage{
trans_select(.data, ...)

trans_select_(.data, variable)

\method{trans_select_}{LabelDictionary}(.data, variable)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{One or more unquoted translation names separated by commas.}

\item{variable}{A character vector holding the names of the variable translations that
should be picked.}
}
\value{
A new \link[=new_dictionary]{LabelDictionary} class object, holding the picked variable translations.
}
\description{
The functions \code{\link[=trans_select]{trans_select()}} and \code{\link[=trans_select_]{trans_select_()}} pick one or more
variable translations from a \link[=new_dictionary]{LabelDictionary} class object
and create a new \link[=new_dictionary]{LabelDictionary} class object.
The function \code{\link[=trans_select]{trans_select()}} uses non-standard evaluation, whereas
\code{\link[=trans_select_]{trans_select_()}} is the standard evaluation alternative.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=trans_rename]{trans_rename()}}, \code{\link[=trans_set]{trans_set()}},
\code{\link[=trans_merge]{trans_merge()}}, \code{\link[=read_dictionary]{read_dictionary()}}, \code{\link[=write_dictionary]{write_dictionary()}}
}
