% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_set.R
\name{trans_set}
\alias{trans_set}
\alias{trans_set.LabelDictionary}
\alias{trans_set_}
\alias{trans_set_.LabelDictionary}
\title{Change or append a variable translation to an existing \link[=new_dictionary]{LabelDictionary} object}
\usage{
trans_set(.data, ...)

\method{trans_set}{LabelDictionary}(.data, ...)

trans_set_(.data, key, translation)

\method{trans_set}{LabelDictionary}(.data, key, translation)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object}

\item{...}{One or more unquoted expressions separated by commas. Use named
arguments, e.g. ‘new_transation_name = c(a = "A", b = "B")’, to set
translations (named character vectors) to new translation names.}

\item{key}{The name of the variable translation that should be altered.
It can also be variable translation name that does not exist yet.}

\item{translation}{A named character vector holding the new variable
translation that should be assigned to the name given in argument \code{key}.
The names of the character vector \code{translation} correspond to the
original variable values that should be replaced by the new labels.
The values in the character vector \code{translations} are the labels
that should be assigned to the original values.}
}
\value{
An updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
The functions \code{\link[=trans_set]{trans_set()}} and \code{\link[=trans_set_]{trans_set_()}} alter a
\link[=new_dictionary]{LabelDictionary} object. Tjey either alter
or append a translation to a \link[=new_dictionary]{LabelDictionary} object.
The function \code{\link[=trans_set]{trans_set()}} uses named arguments to assign the translations
to the new names (similar to \code{\link[=dplyr:mutate]{dplyr:mutate()}}), whereas the function
\code{\link[=trans_set_]{trans_set_()}} is takes a character string \code{key} holding the
name to which the translation should be assigned and a named character
vector \code{translation} holding the actual translation mapping.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=trans_rename]{trans_rename()}}, \code{\link[=tran_select]{tran_select()}},
\code{\link[=trans_merge]{trans_merge()}}, \code{\link[=read_dictionary]{read_dictionary()}}, \code{\link[=write_dictionary]{write_dictionary()}}
}
