% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate.data.frame}
\alias{translate.data.frame}
\title{Label a variable in a [data.frame]}
\usage{
\method{translate}{data.frame}(df, lexicon, translation,
  col = translation, col_new = col, keep_ordering = FALSE)
}
\arguments{
\item{df}{The data.frame object which contains the variable that should be labelled}

\item{lexicon}{A [LabelLexicon] object, holding various label translations}

\item{translation}{A character string holding the name of the label translation that
should be used for labelling the variable. This name must be one of the translation names
returned by \code{names(lexicon)}.}
}
\value{
An extended data.frame, that has a factor variable holding the
assigned labels.
}
\description{
This function takes a data set and converts one of its categorical variables
(not necessarely a factor variable) to a factor variable with new labels. 
The label assignment is given by a label translation (defined by the arguments
\code{lexicon} and \code{translation}).
}
