% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Relabel a variable in a [data.frame]}
\usage{
translate(...)
}
\arguments{
\item{...}{Various arguents}
}
\value{
An extended data.frame, that has a factor variable holding the
}
\description{
This function takes a data.frame and converts one of its categorical variables
(not necessarely a factor variable) to a factor variable with new labels. 
The label assignment is given by a lexicon entry (defined by the arguments
\code{lexicon} and \code{translation}).
}
