% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_translate.R
\name{translate_df}
\alias{translate_df}
\title{This function relabels several variables in a data.frame}
\usage{
translate_df(.data, dictionary, translation, col, col_new, keep_order,
  err_handler)
}
\arguments{
\item{.data}{Either a data frame, a factor or an atomic vector.}

\item{dictionary}{A \link[=new_lama_dictionary]{lama_dictionary} object, holding the translations for various
variables.}

\item{translation}{A character vector holding the names of the variable
translations which
should be used for assigning new labels to the variable. This names must be
a subset of the translation names returned by \code{names(dictionary)}.}

\item{col}{Only used if \code{.data} is a data frame. The argument \code{col} must be
a character vector of the same length as \code{translation} holding
the names of the data.frame columns that
should be relabeled. If omitted, then it will be assumed that the column
names are the same as the given translation names in the argument \code{translation}.}

\item{col_new}{Only used if \code{.data} is a data frame. The argument \code{col} must be
a character vector of the same length as \code{translation} holding
the names under which the relabeled variables should be stored in
the data.frame. If omitted, then it will be assumed that the new column
names are the same as the column names of the original variables.}

\item{keep_order}{A boolean vector of length one or the same length as the
number of translations. If the vector has length one, then the same
configuration is applied to all variable translations. If the vector has
the same length as the number of arguments in \code{...}, then the
to each variable translation there is a corresponding boolean configuration.
If a translated variable in the data.frame is a factor variable,
and the corresponding boolean configuration is set to \code{TRUE}, then the
the order of the original factor variable will be preserved.}

\item{err_handler}{An error handling function}
}
\value{
An factor vector holding the assigned labels.
}
\description{
This function relabels several variables in a data.frame
}
