% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{validate_translation}
\alias{validate_translation}
\title{Check if an object has a valid translation structure}
\usage{
validate_translation(obj,
  err_handler = composerr("The object has not a valid translation structure"))
}
\arguments{
\item{obj}{An object that should be tested}

\item{err_handler}{An error handling function}
}
\description{
This function checks if the object structure is that of a translation (named character vector).
}
