% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dictionary.R
\name{write_dictionary}
\alias{write_dictionary}
\title{Write a \code{yaml} file holding translations for one or multiple variables}
\usage{
write_dictionary(x, yaml_path)
}
\arguments{
\item{x}{A [LabelDictionary] class object holding the variable translations}

\item{yaml_path}{File path, where the yaml file should be saved}
}
\description{
Write a \code{yaml} file holding translations for one or multiple variables
}
