% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml_file.R
\name{write_lexicon_file}
\alias{write_lexicon_file}
\title{Write a \code{yaml} file holding label translations for one or multiple variables}
\usage{
write_lexicon_file(x, yaml_path)
}
\arguments{
\item{x}{A [LabelLexicon] class object holding the labels and translations}

\item{yaml_path}{File path, where the yaml file should be saved}
}
\description{
Write a \code{yaml} file holding label translations for one or multiple variables
}
