% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{yaml_to_dictionary}
\alias{yaml_to_dictionary}
\title{Transform data structure from yaml format to the [LabelDictionary] class input format}
\usage{
yaml_to_dictionary(data)
}
\description{
When a yaml file is read in, the data has the structure
vars (named list) > translations (named list)
This structure is transformed to the [LabelDictionary] class input structure
vars (named list) >  translations (named character vector)
}
