% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{yaml_to_lexicon}
\alias{yaml_to_lexicon}
\title{Transform data structure from yaml format to the [LabelLexicon] class input format}
\usage{
yaml_to_lexicon(data)
}
\description{
When a yaml file is read in the data has the structure
vars (named list) > labels (named list)
This structure is transformed to the [LabelLexicon] class input structure
vars (named list) >  labels (named character vector)
}
