% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dictionary_to_yaml}
\alias{dictionary_to_yaml}
\title{Transform data structure from \link[=new_dictionary]{LamaDictionary} class input format to the yaml format}
\usage{
dictionary_to_yaml(data)
}
\arguments{
\item{data}{A pre-dictionary object.}
}
\value{
An object similar to a pre-dictionary object, but each translation is not a named character vector, but a named list holding character strings.
}
\description{
In the \link[=new_dictionary]{LamaDictionary} class object the data has the structure
vars (named list) > translations (named character vector)
This structure is transformed to the yaml file structure
vars (named list) > translations (named list)
}
