% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_dictionary.R
\name{is.dictionary}
\alias{is.dictionary}
\title{Check if an object is a \link[=new_dictionary]{lama_dictionary} class object}
\usage{
is.dictionary(obj)
}
\arguments{
\item{obj}{The object in question}
}
\value{
\code{TRUE} if the object is a
\link[=new_dictionary]{lama_dictionary} class object, \code{FALSE} otherwise.
}
\description{
Check if an object is a \link[=new_dictionary]{lama_dictionary} class object
}
\seealso{
\code{\link[=validate_dictionary]{validate_dictionary()}}, \code{\link[=as.dictionary]{as.dictionary()}}, \code{\link[=new_dictionary]{new_dictionary()}},
\code{\link[=lama_translate]{lama_translate()}}, \code{\link[=lama_read]{lama_read()}}, \code{\link[=lama_write]{lama_write()}}, \code{\link[=lama_select]{lama_select()}},
\code{\link[=lama_rename]{lama_rename()}}, \code{\link[=lama_mutate]{lama_mutate()}}, \code{\link[=lama_merge]{lama_merge()}}
}
