% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_rename.R
\name{lama_rename}
\alias{lama_rename}
\alias{lama_rename_}
\alias{lama_rename_.LabelDictionary}
\title{Rename multiple variable translations in a \link[=new_dictionary]{LabelDictionary} object}
\usage{
lama_rename(.data, ...)

lama_rename_(.data, old, new)

\method{lama_rename_}{LabelDictionary}(.data, old, new)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{One or more unquoted expressions separated by commas. Use named arguments, e.g. ‘new_name = old_name’, to rename selected variables.}

\item{old}{A character vector holding the names of the variable translations,
that should be renamed.}

\item{new}{A character vector holding the new names of the variable
translations.}
}
\value{
The updated \link[=new_dictionary]{LabelDictionary} class object.
}
\description{
The functions \code{\link[=lama_rename]{lama_rename()}} and \code{\link[=lama_rename_]{lama_rename_()}}
are used to rename one or more variable translations inside of a
\link[=new_dictionary]{LabelDictionary} class object.
The function \code{\link[=lama_rename]{lama_rename()}} uses non-standard evaluation,
whereas \code{\link[=lama_rename_]{lama_rename_()}} is the standard evaluation alternativ.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=lama_select]{lama_select()}}, \code{\link[=lama_set]{lama_set()}},
\code{\link[=lama_merge]{lama_merge()}}, \code{\link[=lama_read]{lama_read()}}, \code{\link[=lama_write]{lama_write()}}
}
