% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_select.R
\name{lama_select}
\alias{lama_select}
\alias{lama_select_}
\alias{lama_select_.LabelDictionary}
\title{Select multiple variable translations and create a new \link[=new_dictionary]{LabelDictionary} object}
\usage{
lama_select(.data, ...)

lama_select_(.data, key)

\method{lama_select_}{LabelDictionary}(.data, key)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{One or more unquoted translation names separated by commas.}

\item{key}{A character vector holding the names of the variable translations that
should be picked.}
}
\value{
A new \link[=new_dictionary]{LabelDictionary} class object, holding the picked variable translations.
}
\description{
The functions \code{\link[=lama_select]{lama_select()}} and \code{\link[=lama_select_]{lama_select_()}} pick one or more
variable translations from a \link[=new_dictionary]{LabelDictionary} class object
and create a new \link[=new_dictionary]{LabelDictionary} class object.
The function \code{\link[=lama_select]{lama_select()}} uses non-standard evaluation, whereas
\code{\link[=lama_select_]{lama_select_()}} is the standard evaluation alternative.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=lama_rename]{lama_rename()}}, \code{\link[=lama_set]{lama_set()}},
\code{\link[=lama_merge]{lama_merge()}}, \code{\link[=lama_read]{lama_read()}}, \code{\link[=lama_write]{lama_write()}}
}
