% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{mutate}
\alias{mutate}
\alias{mutate.LabelDictionary}
\title{Change a variable translation or append a variable translation to an existing [LabelDictionary] object}
\usage{
mutate(.data, ...)

\method{mutate}{LabelDictionary}(.data, variable, translation)
}
\arguments{
\item{.data}{A [LabelDictionary] object}

\item{...}{Various arguments}

\item{variable}{The name of the variable translation that should be altered.
If can also be variable translation name that does not exist yet.}

\item{translation}{A named character vector holding the new variable
translation that should be assigned to the name given in argument \code{variable}.
The names of the character vector \code{translation} correspond to the
original variable values that should be replaced by the new labels.
The values in the character vector \code{translations} are the labels
that should be assigned to the original values.}
}
\value{
An updated [LabelDictionary] class object.
}
\description{
This function alters a [LabelDictionary] object. It can be used for altering
or appending a translation to a [LabelDictionary] object.
}
