% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{rename}
\alias{rename}
\alias{rename.LabelDictionary}
\title{Rename multiple variable translations in a [LabelDictionary] object}
\usage{
rename(.data, ...)

\method{rename}{LabelDictionary}(.data, old, new)
}
\arguments{
\item{.data}{A [LabelDictionary] object, holding the variable translations}

\item{...}{Various arguments}

\item{old}{A character vector holding the names of the variable translations, that should be renamed.}

\item{new}{A character vector holding the new names of the variable translations.}
}
\value{
The updated [LabelDictionary] class object.
}
\description{
This function renamess one or more variable translations inside of a [LabelDictionary] class object.
}
