% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{select}
\alias{select}
\alias{select.LabelDictionary}
\title{Pick multiple variable translations and create a new \link[=new_dictionary]{LabelDictionary} object}
\usage{
select(.data, ...)

\method{select}{LabelDictionary}(.data, variable)
}
\arguments{
\item{.data}{A \link[=new_dictionary]{LabelDictionary} object, holding the variable translations}

\item{...}{Various arguments}

\item{variable}{A character vector holding the names of the variable translations that
should be picked.}
}
\value{
A new \link[=new_dictionary]{LabelDictionary} class object, holding the picked variable translations.
}
\description{
This function picks one or more variable translations from a \link[=new_dictionary]{LabelDictionary} class object
and creates a new \link[=new_dictionary]{LabelDictionary} class object.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link[=new_dictionary]{new_dictionary()}}, \code{\link[=rename]{rename()}}, \code{\link[=mutate]{mutate()}}, \code{\link[=merge]{merge()}}, \code{\link[=read_directory]{read_directory()}}, \code{\link[=write_directory]{write_directory()}}
}
