% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stringify}
\alias{stringify}
\title{Coerce a vector into a character string (\code{'x1', 'x2', ...})}
\usage{
stringify(x)
}
\arguments{
\item{x}{A vector that should be coerced.}
}
\value{
A character string holding the collapsed vector.
}
\description{
Coerce a vector into a character string (\code{'x1', 'x2', ...})
}
