% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lama_translate.R
\name{translate_variable}
\alias{translate_variable}
\title{This function relabels a vector}
\usage{
translate_variable(val, translation, keep_order, err_handler)
}
\arguments{
\item{val}{The vector that should be relabeled. Allowed are all vector types (also factor).}

\item{translation}{Named character vector holding the label assignments.}

\item{keep_order}{A logical flag. If the vector in \code{val}
is a factor variable and \code{keep_order} is set to \code{TRUE}, then
the order of the original factor variable is preserved.}

\item{err_handler}{An error handling function}
}
\value{
A factor vector holding the assigned labels
}
\description{
This function relabels a vector
}
