% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{yaml_to_dictionary}
\alias{yaml_to_dictionary}
\title{Transform data structure from yaml format to the \link[=new_lama_dictionary]{lama_dictionary} class input format}
\usage{
yaml_to_dictionary(data)
}
\arguments{
\item{data}{An object similar to a lama-dictionary object, but each translation is not a named character vector, but a named list holding character strings.}
}
\value{
A list that has lama-dictionary structure.
}
\description{
When a yaml file is read in, the data has the structure
vars (named list) > translations (named list)
This structure is transformed to the \link[=new_lama_dictionary]{lama_dictionary} class input structure
vars (named list) >  translations (named character vector)
}
